+++ CMineNew/Src/Map/BlockData/Static/BlockStaticDataTexturedCross.cs
@@ -4,14 +4,14 @@ namespace CMineNew.Map.BlockData.Static{
     public class BlockStaticDataTexturedCross : BlockStaticDataCross{
         private readonly Area2d _textureArea;
 
-        public BlockStaticDataTexturedCross(string id, bool passable, bool lightSource, int lightSourceLight,
-            int blockLightPassReduction, int sunlightPassReduction, string texture)
+        public BlockStaticDataTexturedCross(string id, bool passable, bool lightSource, sbyte lightSourceLight,
+            sbyte blockLightPassReduction, sbyte sunlightPassReduction, string texture)
             : base(id, passable, lightSource, lightSourceLight, blockLightPassReduction, sunlightPassReduction) {
             _textureArea = CMine.Textures.Areas[texture];
         }
 
-        public BlockStaticDataTexturedCross(string id, bool passable, bool lightSource, int lightSourceLight,
-            int blockLightPassReduction, int sunlightPassReduction, Area2d texture)
+        public BlockStaticDataTexturedCross(string id, bool passable, bool lightSource, sbyte lightSourceLight,
+            sbyte blockLightPassReduction, sbyte sunlightPassReduction, Area2d texture)
             : base(id, passable, lightSource, lightSourceLight, blockLightPassReduction, sunlightPassReduction) {
             _textureArea = texture;
         }
